chatgpt-api/chatgpt-api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.58.1
python-dotenv==1.0.0
